Modified files for single machine
