+++ PeerInfo.cfg
@@ -1,3 +1,3 @@
-1001 <HOSTNAME1> 12311 0
-1002 <HOSTNAME2> 12312 0
-1003 <HOSTNAME3> 12313 0
+1001 Jazzycrazzy 12311 1
+1002 THE_CORE 12312 0
+1003 Anonymous 12313 0
