--- src/bin/.gitignore
@@ -1,19 +1,0 @@
-/BitfieldPayload.class
-/ConfigParser.class
-/FileManager.class
-/FilePiece.class
-/HandShakeMsg.class
-/HavePayload.class
-/Logger.class
-/Message.class
-/Payload.class
-/Peer.class
-/PeerProcess.class
-/PiecePayload.class
-/RequestPayload.class
-/.classpath
-/.gitignore
-/.project
-/PeerProcess$1.class
-/PeerProcess$2.class
-/log_peer_123.log
